amandaye_backend/requirements.txt
Django==4.2.21
mysqlclient==2.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1